Add middleware to app level, remove overuse case of try catch, better Error handling

+++ src/controllers/book.ts
@@ -1,77 +1,72 @@
 // Dependencies
-import { Request, Response } from 'express';
 import { IBookQuery } from '../interfaces/book-query';
 
-// Models
+// Models, Interface
+import { Request, Response } from 'express';
 import { Book } from '../models/book';
 
 // Services
-import { verifyUserService } from '../services/verify-user';
 import { getBookService } from '../services/get-book';
+import { validateEditFieldService } from '../services/validate-edit-field';
 
 // Constant/Enum
 import {
   BOOK_NON_STRING_QUERIES,
-  BOOK_PATCHABLE_FIELDS,
   COMMON_QUERIES
 } from '../utils/common/book-filters';
 
 // Utils
 import { toLowerCaseQuery } from '../utils/functions/lower-case-query';
-import { CustomError } from '../utils/classes/custom-error-class';
 import { SortOrder } from 'mongoose';
+import { UserRoleEnum } from '../utils/common/enum';
+
+// Error handling
+import CustomError from '../errors/custom-errors';
 
 class BookController {
-  public async createNewBook(req: Request, res: Response): Promise<void> {
-    try {
-      await verifyUserService.isGuest(req, res);
-      const book = await new Book({
-        ...req.body,
-        user_id: req.user?._id
-      }).save();
-
-      res.status(201).send({
-        success: true,
-        data: { message: 'New book created successfully', book }
-      });
-    } catch (error: any) {
-      res
-        .status(error.errorCode | 400)
-        .send({ success: false, message: error.message });
+  public async createNewBook(
+    req: Request,
+    res: Response
+  ): Promise<Response | void> {
+    if (req.userRole === UserRoleEnum.GUEST) {
+      throw new CustomError('Please authenticate', 401);
     }
+    const book = await new Book({
+      ...req.body,
+      user_id: req.user?._id
+    }).save();
+
+    res.status(201).send({
+      success: true,
+      data: { message: 'New book created successfully', book }
+    });
   }
 
-  public async editSpecificBook(req: Request, res: Response): Promise<void> {
-    const isValidEdition = Object.keys(req.body).every((k) =>
-      BOOK_PATCHABLE_FIELDS.includes(k)
-    );
-    if (!isValidEdition) {
-      throw new CustomError('Operation Not Valid');
+  public async editSpecificBook(
+    req: Request,
+    res: Response
+  ): Promise<Response | void> {
+    if (req.userRole === UserRoleEnum.GUEST) {
+      throw new CustomError('Please authenticate', 401);
     }
-    try {
-      await verifyUserService.isGuest(req, res);
-      const book = await Book.findOneAndUpdate(
-        {
-          _id: req.params.id,
-          user_id: req.user?._id
-        },
-        req.body,
-        { new: true }
-      );
+    validateEditFieldService.isValidateBookEditFields(req, res);
 
-      if (!book) {
-        throw new CustomError('No Book Found', 404);
-      }
-      res.send({ success: true, data: { book } });
-    } catch (error: any) {
-      res
-        .status(error.errorCode | 400)
-        .send({ success: false, message: error.message });
+    const book = await Book.findOneAndUpdate(
+      {
+        _id: req.params.id,
+        user_id: req.user?._id
+      },
+      req.body,
+      { new: true }
+    );
+
+    if (!book) {
+      throw new CustomError('No Book Found', 404);
     }
+    res.send({ success: true, data: { book } });
   }
 
   public async getBooks(req: Request, res: Response): Promise<void> {
-    const userRole = await verifyUserService.checkUserRole(req, res);
     const filters: IBookQuery = {};
     const sort: { [key: string]: SortOrder } | undefined = {};
     // Update sort and filters from query
@@ -91,54 +86,45 @@ class BookController {
       }
     });
 
-    try {
-      const books = await getBookService.getBookList(
-        req,
-        res,
-        filters,
-        sort,
-        userRole
-      );
-      res.send({ success: true, data: { books } });
-    } catch (error: any) {
-      res
-        .status(error.errorCode | 400)
-        .send({ success: false, message: error.message });
+    const books = await getBookService.getBookList(
+      req,
+      res,
+      filters,
+      sort,
+      req.userRole as string
+    );
+    if (!books) {
+      throw new CustomError('No Book Found', 404);
     }
+
+    res.send({ success: true, data: { books } });
   }
 
   public async getSpecificBook(req: Request, res: Response): Promise<void> {
-    const userRole = await verifyUserService.checkUserRole(req, res);
-    try {
-      const book = await getBookService.getSpecificBook(req, res, userRole);
-      if (!book) {
-        throw new CustomError('No Book Found', 404);
-      }
-      res.send({ success: true, data: { book } });
-    } catch (error: any) {
-      res
-        .status(error.errorCode | 400)
-        .send({ success: false, message: error.message });
+    const book = await getBookService.getSpecificBook(
+      req,
+      res,
+      req.userRole as string
+    );
+    if (!book) {
+      throw new CustomError('No Book Found', 404);
     }
+    res.send({ success: true, data: { book } });
   }
 
   public async deleteSpecificBook(req: Request, res: Response): Promise<void> {
-    try {
-      await verifyUserService.isGuest(req, res);
-      const book = await Book.findByIdAndDelete({
-        _id: req.params.id,
-        user_id: req.user?._id
-      });
-
-      if (!book) {
-        throw new CustomError('No Book Found', 404);
-      }
-      res.send({ success: true, message: 'Delete successfully' });
-    } catch (error: any) {
-      res
-        .status(error.errorCode | 400)
-        .send({ success: false, message: error.message });
+    if (req.userRole === UserRoleEnum.GUEST) {
+      throw new CustomError('Please authenticate', 401);
     }
+    const book = await Book.findByIdAndDelete({
+      _id: req.params.id,
+      user_id: req.user?._id
+    });
+
+    if (!book) {
+      throw new CustomError('No Book Found', 404);
+    }
+    res.send({ success: true, message: 'Delete successfully' });
   }
 }
